Fix ServiceManager class. Move ServiceManagerSupplements into logic level.
