+++ CarRental.BussinessLayer/Managers/ServiceManagerSupplements.cs
@@ -3,14 +3,17 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using CarRental.Models.Car.RecordTypes;
+using CarRental.Data.Models.Car;
+using CarRental.Data.Models.Car.RecordTypes;
+using CarRental.BussinessLayer.Managers;
+using CarRental.BussinessLayer.Interfaces;
 
 // HILLEL, C# PRO COURSE, TEACHER: MARIIA DZIVINSKA
 // HOMEWORK: "ДЗ 4. Extension methods & Record type."
 // STUDENT: PARKHOMENKO YAROSLAV
 // DATE: 03-JUN-2024
 
-namespace CarRental.Data.Models.Car;
+namespace CarRental.BussinessLayer.Managers;
 
 public class ServiceManagerSupplements
 {
